download service refactor. add es query for sequence file data
