+++ .env.example
@@ -21,6 +21,7 @@ VAULT_TOKEN=
 ES_HOST=http://localhost:9200
 ES_USER=''
 ES_PASS=''
+ES_INDEX=''
 
 ###### 
 PORT=4000
